# sivusto/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/quizzes": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "List Quizzes",
        "operationId": "list_quizzes",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuizResponse"
                  },
                  "type": "array",
                  "title": "Response List Quizzes"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Create Quiz",
        "operationId": "create_quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/quizzes/{quiz_id}": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "Get Quiz",
        "operationId": "get_quiz",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Quiz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "quizzes"
        ],
        "summary": "Update Quiz",
        "operationId": "update_quiz",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Quiz Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quizzes"
        ],
        "summary": "Delete Quiz",
        "operationId": "delete_quiz",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Quiz Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/quiz-attempts/{attempt_id}/responses": {
      "get": {
        "tags": [
          "quiz_responses"
        ],
        "summary": "List Responses",
        "operationId": "list_responses",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuizResponseItemSchema"
                  },
                  "title": "Response List Responses"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "quiz_responses"
        ],
        "summary": "Create Response",
        "operationId": "create_response",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizResponseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponseItemSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quiz-responses/{response_id}": {
      "get": {
        "tags": [
          "quiz_responses"
        ],
        "summary": "Get Response",
        "operationId": "get_response",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponseItemSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "quiz_responses"
        ],
        "summary": "Delete Response",
        "operationId": "delete_response",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuizCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionCreate"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "title": "QuizCreate"
      },
      "QuizQuestionCreate": {
        "properties": {
          "question_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text"
          },
          "iala_light_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iala Light Id"
          },
          "position": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": -2147483648.0,
            "title": "Position",
            "default": 0
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionOptionCreate"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "title": "QuizQuestionCreate"
      },
      "QuizQuestionOptionCreate": {
        "properties": {
          "option_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Text"
          },
          "iala_light_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iala Light Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct",
            "default": false
          },
          "position": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": -2147483648.0,
            "title": "Position",
            "default": 0
          }
        },
        "type": "object",
        "title": "QuizQuestionOptionCreate"
      },
      "QuizQuestionOptionResponse": {
        "properties": {
          "option_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option Text"
          },
          "iala_light_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iala Light Id"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct",
            "default": false
          },
          "position": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": -2147483648.0,
            "title": "Position",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "created_at"
        ],
        "title": "QuizQuestionOptionResponse"
      },
...
      "QuizQuestionResponse": {
        "properties": {
          "question_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Text"
          },
          "iala_light_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iala Light Id"
          },
          "position": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": -2147483648.0,
            "title": "Position",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "quiz_id": {
            "type": "string",
            "format": "uuid",
            "title": "Quiz Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionOptionResponse"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "quiz_id",
          "created_at"
        ],
        "title": "QuizQuestionResponse"
      },
...
      "QuizResponse": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            },
            "type": "array",
            "title": "Questions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "title": "QuizResponse"
      },
      "QuizResponseCreate": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Option Id"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "selected_option_id"
        ],
        "title": "QuizResponseCreate"
      },
      "QuizResponseItemSchema": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "attempt_id": {
            "type": "string",
            "format": "uuid",
            "title": "Attempt Id"
          },
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "selected_option_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Option Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_id",
          "question_id",
          "selected_option_id",
          "created_at"
        ],
        "title": "QuizResponseItemSchema"
      },
      "QuizUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "QuizUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}